Update Tailwind to v4
When following along with the tutorial, I hit a snag when getting to the
Tailwind instrucitons. With the version bump from v3 to v4 Tailwind
moved do a different installation and setup process, which means that
when running

   npm i -D tailwindcss

we'll get version 4 which doesn't work with the instructions as written.

This commit updates both the instructions and the included code to the
latest version of Tailwind, including the modifications made to the
typography plugin.